Added Dithering
It actually dithers correctly using lightness calculations to enhance
the image. It also places it in the world mostly correct. There may be
some issues regarding lighting, unloaded chunks with horizontal images,
and minor coloration issues which isn't that big of a deal. This should
be considered the first "Release"